DataAPI/requirements.txt
Flask==3.0.2
requests==2.31.0
python-memcached==1.62

